store-availability-api/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
httpx==0.27.0
httpcore==1.0.5
google-generativeai==0.7.2